Remove workaround for MSVC not detecting [[noreturn]]

+++ src/libmettle/windows/subprocess.cpp
@@ -51,7 +51,7 @@ namespace mettle::windows {
 
     while(true) {
       auto result = WaitForMultipleObjects(
-        total_size, events.get(), false, timeout
+        static_cast<DWORD>(total_size), events.get(), false, timeout
       );
       if(result == WAIT_FAILED || result == WAIT_TIMEOUT) {
         cancel_io(dests);
